+++ compiler/AST/alist.cpp
@@ -25,61 +25,6 @@
 #include "stmt.h"
 #include "stringutil.h"
 
-AList::AList() :
-  head(NULL),
-  tail(NULL),
-  parent(NULL),
-  length(0)
-{ }
-
-
-Expr* AList::first(void) {
-  return head;
-}
-
-
-Expr* AList::last(void) {
-  return tail;
-}
-
-
-Expr* AList::only(void) {
-  if (!head)
-    INT_FATAL("only() called on empty list");
-
-  if (head == tail)
-    return first();
-  else
-    INT_FATAL("only() called on list with more than one element");
-
-  return NULL;
-}
-
-
-Expr* AList::get(int index) const {
-  if (index <=0) {
-    INT_FATAL("Indexing list must use positive integer");
-  }
-
-  int i = 0;
-
-  for_alist(node, *this) {
-    i++;
-
-    if (i == index) {
-      return node;
-    }
-  }
-
-  INT_FATAL("Indexing list out of bounds");
-
-  return NULL;
-}
-
-bool AList::empty() const {
-  return length == 0;
-}
-
 void AList::insertAtHead(Expr* new_ast) {
   if (new_ast->parentSymbol || new_ast->parentExpr)
     INT_FATAL(new_ast, "Argument is already in AST in AList::insertAtHead");
