+++ compiler/AST/symbol.cpp
@@ -232,53 +232,11 @@ bool Symbol::isRenameable() const {
   return true;
 }
 
-bool Symbol::isRef() {
-  QualifiedType q = qualType();
-  return (type != NULL) && (q.isRef() || type->symbol->hasFlag(FLAG_REF));
-}
-
-bool Symbol::isWideRef() {
-  QualifiedType q = qualType();
-  return (q.isWideRef() || type->symbol->hasFlag(FLAG_WIDE_REF));
-}
-
-bool Symbol::isRefOrWideRef() {
-  return isRef() || isWideRef();
-}
-
-
 // Returns the scope in which the given symbol is declared; NULL otherwise.
 BlockStmt* Symbol::getDeclarationScope() const {
   return (defPoint != NULL) ? defPoint->getScopeBlock() : NULL;
 }
 
-
-bool Symbol::hasFlag(Flag flag) const {
-  CHECK_FLAG(flag);
-  return flags[flag];
-}
-
-
-void Symbol::addFlag(Flag flag) {
-  CHECK_FLAG(flag);
-  flags.set(flag);
-}
-
-
-void Symbol::copyFlags(const Symbol* other) {
-  flags |= other->flags;
-  qual   = other->qual;
-}
-
-
-void Symbol::removeFlag(Flag flag) {
-  CHECK_FLAG(flag);
-  flags.reset(flag);
-}
-
-bool Symbol::hasEitherFlag(Flag aflag, Flag bflag) const {
-  return hasFlag(aflag) || hasFlag(bflag);
-}
 
 bool Symbol::isKnownToBeGeneric() {
   if (FnSymbol* fn = toFnSymbol(this))
